moved old files intonew structure added klassen rümpfe
