--- game/src/MainMenu/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod main_menu;
-pub(crate) mod Button;
