+++ game/src/backend/mod.rs
@@ -1,1 +1,7 @@
 pub(crate) mod asset_service;
+pub(crate) mod error;
+pub(crate) mod gamestate;
+pub(crate) mod movement;
+pub(crate) mod screen;
+pub(crate) mod utils;
+
