+++ game/src/backend/screen.rs
@@ -1,6 +1,6 @@
 use crate::error::RLError;
-use crate::mainmenu::{MainMenu, Message};
-use crate::utils::get_scale;
+use crate::main_menu::mainmenu::{MainMenu, Message};
+use crate::backend::utils::get_scale;
 use crate::RLResult;
 use ggez::graphics::Color;
 use ggez::{event, graphics, Context};
