+++ game/src/basis/basis.rs
@@ -1,0 +1,14 @@
+use crate::basis::ressourcen::Ressourcen;
+
+struct Basis{
+    absolute:Ressourcen,
+    relative:Ressourcen,
+  //  maschienen:Vec<Maschien> // macht sander maschienen existiert
+
+
+
+
+
+
+}
+
