+++ game/src/basis/ressourcen.rs
@@ -1,0 +1,10 @@
+
+pub(crate) struct Ressourcen{
+    energy:usize,
+    oxygen:usize,
+    live:usize
+}
+
+impl Ressourcen{
+
+}
