+++ game/src/game_core/event.rs
@@ -1,0 +1,3 @@
+pub(crate) struct Event{
+
+}
