+++ game/src/main_menu/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod mainmenu;
+pub(crate) mod button;
+
